feat: Split core in models and services (#47)
* feat: Divide src/core to models and services

* chore: Sort imports

---------
