+++ src/platforms/Instagram.ts
@@ -1,14 +1,15 @@
-import Storage from "../core/Storage";
-import Logger from "../core/Logger";
-import Platform from "../core/Platform";
-import InstagramAuth from "../auth/InstagramAuth";
-import { PlatformId } from ".";
-import Folder from "../core/Folder";
-import Post from "../core/Post";
-import { PostStatus } from "../core/Post";
 import * as fs from "fs";
 import * as path from "path";
 import * as sharp from "sharp";
+
+import Folder from "../models/Folder";
+import InstagramAuth from "../auth/InstagramAuth";
+import Logger from "../services/Logger";
+import Platform from "../models/Platform";
+import { PlatformId } from ".";
+import Post from "../models/Post";
+import { PostStatus } from "../models/Post";
+import Storage from "../services/Storage";
 
 /**
  * Instagram: support for instagram platform.
