+++ src/services/Fairpost.ts
@@ -1,16 +1,17 @@
 import * as fs from "fs";
 import * as path from "path";
-import Feed from "./Feed";
-import Platform from "./Platform";
-import Storage from "./Storage";
-import Logger from "./Logger";
 import * as platforms from "../platforms";
+
+import Feed from "../models/Feed";
+import Logger from "./Logger";
+import Platform from "../models/Platform";
+import Storage from "./Storage";
 
 /**
  * Fairpost - singleton service for fairpost app
  *
- * - provides getFeed()
- *   all the fairpost logic comes from a feed
+ * - provides the feed, that contains
+ *   all the fairpost logic
  */
 
 class Fairpost {
@@ -56,8 +57,14 @@ class Fairpost {
     return new Feed(this.platforms.filter((p) => p.active));
   }
 
-  public fatal(msg: string) {
+  public error(msg: string) {
+    Logger.error(msg);
+    throw new Error(msg);
+  }
+
+  public fatal(msg: string, code = 1) {
     Logger.fatal(msg);
+    process.exitCode = code;
     throw new Error(msg);
   }
 }
